Start config refactor.
