+++ politeiawww/identity.go
@@ -20,7 +20,12 @@ import (
 	"github.com/decred/politeia/util"
 )
 
-// getIdentity fetches the remote identity from politeiad.
+const (
+	allowInteractive = "i-know-this-is-a-bad-idea"
+)
+
+// getIdentity fetches the remote identity from politeiad and saves it to
+// disk. politeiawww will load it from disk on future startups.
 func getIdentity(rpcHost, rpcCert, rpcIdentityFile, interactive string) error {
 	id, err := remoteIdentity(false, rpcHost, rpcCert)
 	if err != nil {
